backend/requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

typing_extensions>=4.8,<5

torch==2.5.1+cpu
torchvision==0.20.1+cpu

Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
numpy==2.2.6
scikit-learn==1.4.2
opencv-python-headless==4.12.0.88
pillow==11.3.0
psutil==7.0.0
PyMuPDF==1.26.3
huggingface-hub==0.34.2
doclayout_yolo==0.0.4

matplotlib==3.9.2
tqdm>=4.66
pyyaml>=6.0.1

langchain==0.3.27  
langchain-openai==0.3.29  
langchain-google-genai==2.0.10  
langchain-community==0.3.27  
google-generativeai==0.8.5 
requests>=2.25.0
google-cloud-texttospeech==2.27.0
pydub>=0.25.0


google-genai==1.2.0
azure-cognitiveservices-speech==1.45.0